E2E fixes
